Reorganize code, add test for new code
